pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fastapi-alertengine"
version = "1.6.0"
description = "FastAPI AlertEngine: production-ready request monitoring with p95 latency, error rate, Slack alerts, and zero-config instrument() setup."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Tandem Media" }
]
license = "MIT"
dependencies = [
  "fastapi",
  "redis",
  "pydantic-settings",
  "httpx",
  "PyJWT",
]

[project.optional-dependencies]
dev = [
  "uvicorn[standard]",
  "pytest",
  "pytest-asyncio",
  "httpx",
  "fakeredis",
  "anyio",
]

[project.urls]
Homepage = "https://github.com/Tandem-Media/fastapi-alertengine"
